# Ficha 8/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Ficha 7 API",
        "description": "Ficha 7 API Information",
        "contact": {
            "name": "TPSI-DWB"
        },
        "servers": [
            "http://localhost:3000"
        ]
    },
    "paths": {
        "/person": {           
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "Get persons",
                "description": "Get all persons.",
                "operationId": "getPersons",
                "produces": [
                    "application/xml",
                    "application/json"
                ],               
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "No persons found"
                    }
                }
            }
        },
        "/person/{id}": {
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "Get person by id",
                "description": "Get person by ID.",
                "operationId": "getPersonById",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id that needs to be fetched. Use id 1 for testing. ",
                        "required": true,
                        "type": "string"
                    }                    
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }         
        }
    },
    "definitions": {
        "Person": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-primary-key": true
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "age": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "xml": {
                "name": "Person"
            }
        }
    }
}